add domain description tests
