+++ test/domain_description/domain_wrong_absorber_type.ini
@@ -1,0 +1,5 @@
+[ABSORBER]
+type = Amplifier
+n_lambda_depth = 3.
+round_trip = 1e-25
+degree = 5
