+++ test/domain_description/test_file_input.py
@@ -1,19 +1,23 @@
+import configparser
 import pathlib
 
 import numpy as np
 import pytest
 
 import hsc.domain_properties as d
+from hsc.domain_properties.file_input import read_absorber_config
 
-templates_path = pathlib.Path.cwd().joinpath("templates")
 
-
-def test_file_input_domain_general():
+def test_file_input_domain_general(templates_path):
     for file in ["domain_c_shaped.ini", "domain_cylindrical.ini", "domain_none.ini"]:
         template = templates_path.joinpath(file)
         des = d.read_config(template)[0]
 
-        assert all(np.in1d(des.frequencies, np.linspace(4000, 20000, 10))[: len(des.frequencies)])
+        assert all(
+            np.in1d(des.frequencies, np.linspace(4000, 20000, 10))[
+                : len(des.frequencies)
+            ]
+        )
         assert des.rho == 1.2
         assert des.c == 343.0
         c = des.crystal
@@ -21,7 +25,7 @@ def test_file_input_domain_general():
         assert c.n == 10
 
 
-def test_file_input_domain_c_shaped():
+def test_file_input_domain_c_shaped(templates_path):
     template = templates_path.joinpath("domain_c_shaped.ini")
     descriptions = d.read_config(template)
     crystals = [des.crystal for des in descriptions]
@@ -42,7 +46,7 @@ def test_file_input_domain_c_shaped():
     assert set(gap_ws) == set(np.outer(inner_rs, np.linspace(0.3, 1.0, 2)).flatten())
 
 
-def test_file_input_domain_cylindrical():
+def test_file_input_domain_cylindrical(templates_path):
     template = templates_path.joinpath("domain_cylindrical.ini")
     descriptions = d.read_config(template)
     crystals = [des.crystal for des in descriptions]
@@ -56,7 +60,7 @@ def test_file_input_domain_cylindrical():
     assert np.allclose(rs, np.linspace(6.5e-3, 10e-3, 5))
 
 
-def test_file_input_domain_none():
+def test_file_input_domain_none(templates_path):
     template = templates_path.joinpath("domain_none.ini")
     descriptions = d.read_config(template)
     crystals = [des.crystal for des in descriptions]
@@ -68,3 +72,11 @@ def test_unknown_crystal_type():
     template = pathlib.Path(__file__).parent.joinpath("domain_wrong_crystal_type.ini")
     with pytest.raises(TypeError):
         d.read_config(template)
+
+
+def test_wrong_absorber():
+    template = pathlib.Path(__file__).parent.joinpath("domain_wrong_absorber_type.ini")
+    config = configparser.ConfigParser()
+    config.read(template)
+    with pytest.raises(TypeError):
+        read_absorber_config(config, 1e-3)
